requirements.txt
# Traceset — pinned dependencies

# ─── Web framework ────────────────────────────────────────────────────────────
fastapi==0.115.12
uvicorn[standard]==0.34.2

# ─── Config & validation ──────────────────────────────────────────────────────
pydantic==2.11.4
pydantic-settings==2.9.1

# ─── Database ─────────────────────────────────────────────────────────────────
supabase==2.15.2

# ─── Sentinel: GitHub ─────────────────────────────────────────────────────────
PyGitHub==2.9.1
PyNaCl==1.5.0

# ─── Sentinel: Papers ─────────────────────────────────────────────────────────
arxiv==3.0.0
feedparser==6.0.11
semanticscholar==0.12.0

# ─── Sentinel: Patents (USPTO PatentsView REST API — no scraping) ─────────────
httpx==0.28.1

# ─── AI / Claude ──────────────────────────────────────────────────────────────
anthropic==0.49.0

# ─── Dashboard ────────────────────────────────────────────────────────────────
streamlit==1.45.1

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.3.5
pytest-asyncio==0.26.0
respx==0.22.0
